backend done

+++ src/main/java/com/project/noteapp/services/NoteServicesImplementation.java
@@ -4,7 +4,6 @@ import com.project.noteapp.model.Note;
 import com.project.noteapp.model.User;
 import com.project.noteapp.repository.NoteRepository;
 import com.project.noteapp.repository.UserRepository;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,7 +22,7 @@ public class NoteServicesImplementation implements NoteServices {
     @Override
     @Transactional
     public void newNote(Integer id, Note note) {
-        User user = userRepository.findById(id).get();
+        User user = userRepository.findById(id).orElseThrow();
         note.setUser(user);
         note.setCreatedDate(LocalDate.now());
         note.setEditedDate(LocalDate.now());
@@ -32,27 +31,32 @@ public class NoteServicesImplementation implements NoteServices {
 
     @Override
     @Transactional
-    public void editNote(Integer note_id, Note note) {
-        note.setNoteId(note_id);
-        note.setEditedDate(LocalDate.now());
-        noteRepository.save(note);
+    public boolean editNote(Integer note_id, Note note, Integer userId) {
+        Note note1 = noteRepository.findById(note_id).orElseThrow();
+        if (note1.getUser().getUserId().equals(userId)) {
+            note.setNoteId(note_id);
+            note.setEditedDate(LocalDate.now());
+            noteRepository.save(note);
+            return true;
+        } else return false;
     }
 
     @Override
     @Transactional
-    public boolean deleteNote(Integer id) {
-        try {
+    public boolean deleteNote(Integer id, Integer userId) {
+        Note note = noteRepository.findById(id).orElseThrow();
+        if (note.getUser().getUserId().equals(userId)) {
             noteRepository.deleteById(id);
             return true;
-        } catch (Exception e) {
-            return false;
-        }
+        } else return false;
     }
 
     @Override
     @Transactional(readOnly = true)
-    public Note getNoteById(Integer id) {
-        return noteRepository.findById(id).get();
+    public Note getNoteById(Integer id, Integer noteId) {
+        return noteRepository.findById(noteId)
+                .filter(note1 -> note1.getUser().getUserId().equals(id))
+                .orElseThrow();
     }
 
     @Override
